requirements.txt
Flask==2.2.0
google-cloud-bigquery==3.10.0
